feat: update role id data type

+++ backend/database/migrations/2014_10_12_000000_create_users_table.php
@@ -16,10 +16,10 @@ return new class extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->tinyInteger('role_id');
+            $table->unsignedBigInteger('role_id');
             $table->rememberToken();
             $table->timestamps();
-            
+
             $table->foreign('role_id')->references('id')->on('ranks');
         });
     }
